Fix a Type Error

+++ Assets/LeapMotion/Scripts/HandController.cs
@@ -244,9 +244,9 @@ public class HandController : MonoBehaviour {
     if (timeleft <= 0.0f) {
 	  if (frameRateText != null)
 	    frameRateText.text = "Data FPS:" + frame.CurrentFramesPerSecond.ToString ("f2") +
-            Environment.NewLine + "Render FPS:" + (accum/frames).ToString("f2");
+            Environment.NewLine + "Render FPS:" + (accum/(float)frames).ToString("f2");
 	  Debug.LogWarning("Update: data fps:" + frame.CurrentFramesPerSecond.ToString ("f2") +
-                 " render fps:" + (accum/frames).ToString("f2"));
+                 " render fps:" + (accum/(float)frames).ToString("f2"));
       timeleft = UPDATE_INTERVAL;
       accum = 0.0f;
       frames = 0;
